+++ src/main/java/project_pet_backEnd/groomer/PetGroomer/dto/orderby/PGOrderBy.java
@@ -1,0 +1,6 @@
+package project_pet_backEnd.groomer.PetGroomer.dto.orderby;
+
+public enum PGOrderBy {
+    NUM_APPOINTMENTS,
+    PG_NAME
+}
