+++ src/main/java/project_pet_backEnd/groomer/PetGroomer/dto/request/ManagerGetByFunctionIdReq.java
@@ -1,0 +1,16 @@
+package project_pet_backEnd.groomer.PetGroomer.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import javax.validation.constraints.NotBlank;
+
+@Data
+@AllArgsConstructor
+public class ManagerGetByFunctionIdReq {
+    @NotBlank
+    private Integer managerId;
+    @NotBlank
+    private String  managerAccount;
+
+}
