+++ src/main/java/project_pet_backEnd/groomer/PetGroomerCollection/dto/PGCollectionInfo.java
@@ -1,0 +1,14 @@
+package project_pet_backEnd.groomer.PetGroomerCollection.dto;
+
+import lombok.Data;
+
+public class PGCollectionInfo {
+
+    private Integer pgId;
+    private String pgName;
+    private String pgPic;
+    private String porTitle;
+    private String porText;
+    private Data porUpload;
+    private String piPicture;
+}
