+++ src/main/java/project_pet_backEnd/groomer/PetGroomerCollection/dto/PGCollectionManagement.java
@@ -1,0 +1,12 @@
+package project_pet_backEnd.groomer.PetGroomerCollection.dto;
+
+import lombok.Data;
+
+public class PGCollectionManagement {
+
+    private String pgName;
+    private String porTitle;
+    private String porText;
+    private Data proUpload;
+    private String piPicture;
+}
