+++ src/main/java/project_pet_backEnd/groomer/PetGroomerCollection/dto/PGWorkManagement.java
@@ -1,0 +1,10 @@
+package project_pet_backEnd.groomer.PetGroomerCollection.dto;
+
+import lombok.Data;
+
+public class PGWorkManagement {
+    private String porTitle;
+    private String porText;
+    private String piPicture;
+    private Data proUpload;
+}
