+++ src/main/java/project_pet_backEnd/groomer/PetGroomerCollection/vo/PortfolioCollect.java
@@ -1,0 +1,13 @@
+package project_pet_backEnd.groomer.PetGroomerCollection.vo;
+
+import lombok.Data;
+
+@Data
+public class PortfolioCollect {
+    private Integer pcNo;
+    private Integer userId;
+    private Integer porId;
+    private java.util.Date pcCreated;
+
+    // 此處省略建構子、Getter 和 Setter 方法
+}
