--- src/main/java/project_pet_backEnd/groomer/dto/ManagerGetByFunctionIdRequest.java
@@ -1,14 +1,0 @@
-package project_pet_backEnd.groomer.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class ManagerGetByFunctionIdRequest {
-
-    private Integer managerId;
-
-    private String  managerAccount;
-
-}
