--- src/main/java/project_pet_backEnd/groomer/dto/PetGrommerOrderBy.java
@@ -1,6 +1,0 @@
-package project_pet_backEnd.groomer.dto;
-
-public enum PetGrommerOrderBy {
-    NUM_APPOINTMENTS,
-    PG_NAME
-}
