--- src/main/java/project_pet_backEnd/groomer/dto/PetGroomerCollectionInfo.java
@@ -1,14 +1,0 @@
-package project_pet_backEnd.groomer.dto;
-
-import lombok.Data;
-
-public class PetGroomerCollectionInfo {
-
-    private int pgId;
-    private String pgName;
-    private String pgPic;
-    private String porTitle;
-    private String porText;
-    private Data porUpload;
-    private String piPicture;
-}
