--- src/main/java/project_pet_backEnd/groomer/dto/PetGroomerCollectionManagement.java
@@ -1,12 +1,0 @@
-package project_pet_backEnd.groomer.dto;
-
-import lombok.Data;
-
-public class PetGroomerCollectionManagement {
-
-    private String pgName;
-    private String porTitle;
-    private String porText;
-    private Data proUpload;
-    private String piPicture;
-}
