--- src/main/java/project_pet_backEnd/groomer/dto/PetGroomerWorkManagement.java
@@ -1,10 +1,0 @@
-package project_pet_backEnd.groomer.dto;
-
-import lombok.Data;
-
-public class PetGroomerWorkManagement {
-    private String porTitle;
-    private String porText;
-    private String piPicture;
-    private Data proUpload;
-}
