--- src/main/java/project_pet_backEnd/groomer/vo/PictureInfo.java
@@ -1,13 +1,0 @@
-package project_pet_backEnd.groomer.vo;
-
-import lombok.Data;
-
-@Data
-public class PictureInfo {
-    private int piNo;
-    private int porId;
-    private byte[] piPicture;
-    private java.util.Date piDate;//util.date
-
-    // 此處省略建構子、Getter 和 Setter 方法
-}
