--- src/main/java/project_pet_backEnd/groomer/vo/PortfolioCollect.java
@@ -1,13 +1,0 @@
-package project_pet_backEnd.groomer.vo;
-
-import lombok.Data;
-
-@Data
-public class PortfolioCollect {
-    private int pcNo;
-    private int userId;
-    private int porId;
-    private java.util.Date pcCreated;
-
-    // 此處省略建構子、Getter 和 Setter 方法
-}
